Fixed a timer overflow that would cause the UI to become unresponsive after 35 minutes or so

+++ src/burner/shock/core/ostimer.h
@@ -13,7 +13,7 @@ public:
     OSTimer( );
     
     void   Reset( );
-    UINT32 GetElapsedTimeMicroseconds( );
+    UINT64 GetElapsedTimeMicroseconds( );
 };
 
 #endif
